solutions/rws-app/src/requirements.txt
aiosqlite>=0.20.0, <1.0.0
httpx>=0.27.2, <0.28.0
aiohttp>=3.11.11, <4.0.0
python_dotenv>=1.0.1, <2.0.0
azure-identity>=1.19.0, <2.0.0
azure-ai-projects==1.0.0b8
azure-functions>=1.23.0
azure-storage-blob>=12.19.0, <13.0.0
azure-search-documents>=11.4.0
cement==2.10.14
pandas>=2.2.3, <3.0.0
pydantic<3.0.0,>=2.10.0
requests>=2.32.3
pillow>=11.0.0, <12.0.0
openai>=1.76.0
semantic-kernel>=1.29.0
pyodbc>=4.0.39
SQLAlchemy>=2.0.28
